feat: extract buildAgentSystemPrompt + parseModelSpec + PROVIDER_ENV_MAP into @polpo-ai/core
Eliminates cloud/open-source prompt duplication. buildAgentSystemPrompt
is pure logic (identity, responsibilities, tone, personality, hierarchy)
that both self-hosted and cloud import from core.

parseModelSpec handles "provider:model" and "provider/model" parsing.
PROVIDER_ENV_MAP centralizes provider → env var name mapping.

src/adapters/engine.ts now delegates to core for the identity block.
src/llm/pi-client.ts delegates to core for model spec parsing.

Published as @polpo-ai/core@0.3.4
